Refactors the logic behind downloading PDF files from Cloudinary so as not to expose sensitive information

+++ src/Application/Services/Implements/UserService.cs
@@ -10,6 +10,7 @@ using backend.src.Domain.Models;
 using backend.src.Domain.Models.Options;
 using backend.src.Infrastructure.Exceptions;
 using backend.src.Infrastructure.Repositories.Interfaces;
+using CloudinaryDotNet;
 using Mapster;
 using Serilog;
 
@@ -20,6 +21,7 @@ namespace backend.src.Application.Services.Implements
         private readonly IConfiguration _configuration;
         private readonly IUserRepository _userRepository;
         private readonly IFileRepository _fileRepository;
+        private readonly IOfferApplicationRepository _applicationRepository;
         private readonly IEmailService _emailService;
         private readonly ITokenService _tokenService;
         private readonly IFileService _fileService;
@@ -29,6 +31,7 @@ namespace backend.src.Application.Services.Implements
             IConfiguration configuration,
             IUserRepository userRepository,
             IFileRepository fileRepository,
+            IOfferApplicationRepository applicationRepository,
             IVerificationCodeRepository verificationCodeRepository,
             IEmailService emailService,
             ITokenService tokenService,
@@ -38,6 +41,7 @@ namespace backend.src.Application.Services.Implements
             _configuration = configuration;
             _userRepository = userRepository;
             _fileRepository = fileRepository;
+            _applicationRepository = applicationRepository;
             _emailService = emailService;
             _verificationCodeRepository = verificationCodeRepository;
             _tokenService = tokenService;
@@ -1400,6 +1404,7 @@ namespace backend.src.Application.Services.Implements
         /// <exception cref="Exception"></exception>
         public async Task<string> UploadCVByIdAsync(UploadCVDTO uploadCVDTO, int userId)
         {
+            // Validar usuario
             Log.Information("Buscando usuario con la ID: {UserId}", userId);
             User? user = await _userRepository.GetByIdAsync(
                 userId,
@@ -1409,6 +1414,20 @@ namespace backend.src.Application.Services.Implements
             {
                 Log.Error("Usuario no encontrado con la ID: {UserId}", userId);
                 throw new KeyNotFoundException("Usuario no encontrado.");
+            }
+
+            // Validar que el CV pueda ser actualizado
+            bool hasApplicationsPending =
+                await _applicationRepository.HasPendingCvRequiredApplication(user.Id);
+            if (hasApplicationsPending)
+            {
+                Log.Error(
+                    "El usuario con ID: {UserId} tiene postulaciones pendientes que requieren CV. No se puede actualizar el CV.",
+                    userId
+                );
+                throw new InvalidOperationException(
+                    "No se puede actualizar el CV porque tienes postulaciones pendientes que requieren un CV."
+                );
             }
             if (user.CV != null && user.UserType == UserType.Estudiante)
             {
@@ -1425,6 +1444,7 @@ namespace backend.src.Application.Services.Implements
                     );
                     throw new Exception("Error al eliminar el CV existente del estudiante");
                 }
+                await _applicationRepository.MarkCvAsInvalidAsync(user.Id);
             }
             var uploadResult = await _fileService.UploadPDFAsync(uploadCVDTO.CVFile, user);
             if (!uploadResult)
@@ -1436,36 +1456,14 @@ namespace backend.src.Application.Services.Implements
             return "CV del usuario actualizado correctamente";
         }
 
-        /// <summary>
-        /// Descarga el CV de un usuario por su ID.
-        /// </summary>
-        /// <param name="userId">ID del usuario.</param>
-        /// <returns>Datos del CV.</returns>
-        /// <exception cref="KeyNotFoundException"></exception>
-        public async Task<GetCVDTO> DownloadCVByIdAsync(int parsedUserId, int userId)
+        public async Task<(
+            MemoryStream FileStream,
+            string FileName,
+            string ContentType
+        )> DownloadCVByIdAsync(int userId)
         {
+            // Validacion de usuario
             Log.Information("Buscando usuario con la ID: {UserId}", userId);
-
-            // Validacion de usuarios
-            bool requestingUserExists = await _userRepository.ExistsByIdAsync(parsedUserId);
-            if (!requestingUserExists)
-            {
-                Log.Error("Usuario solicitante no encontrado con la ID: {UserId}", parsedUserId);
-                throw new KeyNotFoundException("Usuario solicitante no encontrado.");
-            }
-            bool isApplicant = await _userRepository.CheckRoleAsync(
-                parsedUserId,
-                RoleNames.Applicant
-            );
-            if (isApplicant && parsedUserId != userId)
-            {
-                Log.Warning(
-                    "El usuario solicitante con ID: {RequestingUserId} no tiene permiso para descargar el CV del usuario con ID: {UserId}",
-                    parsedUserId,
-                    userId
-                );
-                throw new UnauthorizedAccessException("No tienes permiso para descargar este CV.");
-            }
             User? user = await _userRepository.GetByIdAsync(
                 userId,
                 new UserQueryOptions { IncludeCV = true }
@@ -1475,13 +1473,70 @@ namespace backend.src.Application.Services.Implements
                 Log.Error("Usuario no encontrado con la ID: {UserId}", userId);
                 throw new KeyNotFoundException("Usuario no encontrado.");
             }
-
-            if (user.CV == null || user.UserType != UserType.Estudiante)
+            if (user.CV == null)
             {
                 Log.Warning("El usuario con ID: {UserId} no tiene un CV para descargar.", userId);
                 throw new KeyNotFoundException("El usuario no tiene un CV para descargar");
             }
-            return user.Adapt<GetCVDTO>();
+
+            // Acceso a el archivo por url firmada
+            string signedUrl = await _fileService.BuildSignedUrlForCVAsync(user.CV.PublicId);
+
+            // Descarga del archivo desde el servidor de hosting
+            using var httpClient = new HttpClient();
+            var response = await httpClient.GetAsync(signedUrl);
+            if (!response.IsSuccessStatusCode)
+            {
+                Log.Error(
+                    "Error al descargar el CV del estudiante con ID: {UserId} desde el servidor de hosting. StatusCode: {StatusCode}",
+                    user.Id,
+                    response.StatusCode
+                );
+                throw new Exception(
+                    "Error al descargar el CV del estudiante desde el servidor de hosting"
+                );
+            }
+            var memoryStream = new MemoryStream();
+            await response.Content.CopyToAsync(memoryStream);
+            memoryStream.Position = 0;
+
+            string fileName = $"CV_{user.FirstName}_{user.LastName}.pdf"
+                .Replace(" ", "_")
+                .Replace("á", "a")
+                .Replace("é", "e")
+                .Replace("í", "i")
+                .Replace("ó", "o")
+                .Replace("ú", "u")
+                .Replace("ñ", "n");
+
+            Log.Information(
+                "CV del estudiante con ID: {UserId} descargado exitosamente desde el servidor de hosting",
+                user.Id
+            );
+            return (memoryStream, fileName, "application/pdf");
+        }
+
+        public async Task<HasCVDTO> CheckCVExistsByIdAsync(int userId)
+        {
+            Log.Information("Buscando usuario con la ID: {UserId}", userId);
+
+            // Validacion de usuarios
+            User? user = await _userRepository.GetByIdAsync(
+                userId,
+                new UserQueryOptions { IncludeCV = true }
+            );
+            if (user == null)
+            {
+                Log.Error("Usuario no encontrado con la ID: {UserId}", userId);
+                throw new KeyNotFoundException("Usuario no encontrado.");
+            }
+            if (user.CV == null)
+            {
+                Log.Warning("El usuario con ID: {UserId} no tiene un CV para descargar.", userId);
+            }
+            Log.Information("CV encontrado para el usuario con ID: {UserId}", userId);
+            HasCVDTO hasCVDTO = new() { HasCV = user.CV != null };
+            return hasCVDTO;
         }
 
         public async Task<string> DeleteCVByIdAsync(int userId)
@@ -1502,6 +1557,19 @@ namespace backend.src.Application.Services.Implements
                 Log.Warning("El usuario con ID: {UserId} no tiene un CV para eliminar.", userId);
                 throw new KeyNotFoundException("El usuario no tiene un CV para eliminar");
             }
+            // Validar capacidad de eliminar el CV
+            bool hasApplicationsPending =
+                await _applicationRepository.HasPendingCvRequiredApplication(user.Id);
+            if (hasApplicationsPending)
+            {
+                Log.Warning(
+                    "El usuario con ID: {UserId} tiene postulaciones pendientes que requieren CV. No se puede eliminar el CV.",
+                    userId
+                );
+                throw new InvalidOperationException(
+                    "No se puede eliminar el CV porque tiene postulaciones pendientes que requieren CV."
+                );
+            }
             // Eliminar CV
             var deleteResult = await _fileRepository.DeleteCVAsync(user.CV.PublicId);
             if (!deleteResult)
@@ -1519,6 +1587,9 @@ namespace backend.src.Application.Services.Implements
                 );
                 throw new Exception("Error al actualizar el usuario después de eliminar el CV");
             }
+            // Update closed applications with the outdated cv flag.
+            await _applicationRepository.MarkCvAsInvalidAsync(user.Id);
+
             return "CV eliminado exitosamente";
         }
         #endregion
